css/cssom/setproperty-null.html

<!doctype html>
<html>
<head>
<link rel="help" href="https://www.w3.org/TR/cssom-1/#dom-cssstyledeclaration-setproperty">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script>
var style = document.body.style;

test(function() {
    style.color = 'white';

    assert_equals(style.color, 'white');
    style.setProperty('color', null);
    assert_equals(style.color, 'red');
}, "Verify that setting a CSS property to null is treated like empty string.");

test(function() {
    style.color = 'white';

    assert_equals(style.color, 'white');
    style.setProperty('color', 'red', null);
    assert_equals(style.color, 'red');
}, "Verify that setting a CSS property priority to null is treated like empty string.");
</script>
</body>
</html>
